requirements.txt
bootstrap-flask
Flask_CKEditor
Flask_Login
Flask-Gravatar
Flask_WTF
WTForms
Werkzeug
Flask~=3.0.3
psycopg2-binary
pygame~=2.6.0
pillow~=10.4.0
pandas~=2.2.2